+++ lightly_studio/src/lightly_studio/core/dataset_query/__init__.py
@@ -1,6 +1,12 @@
 from .boolean_expression import AND, NOT, OR
+from .classification_expression import ClassificationField, ClassificationQuery
 from .dataset_query import DatasetQuery
 from .image_sample_field import ImageSampleField
+from .instance_segmentation_expression import (
+    InstanceSegmentationField,
+    InstanceSegmentationQuery,
+)
+from .object_detection_expression import ObjectDetectionField, ObjectDetectionQuery
 from .order_by import OrderByExpression, OrderByField
 from .video_sample_field import VideoSampleField
 
@@ -8,8 +14,14 @@ __all__ = [
     "AND",
     "NOT",
     "OR",
+    "ClassificationField",
+    "ClassificationQuery",
     "DatasetQuery",
     "ImageSampleField",
+    "InstanceSegmentationField",
+    "InstanceSegmentationQuery",
+    "ObjectDetectionField",
+    "ObjectDetectionQuery",
     "OrderByExpression",
     "OrderByField",
     "VideoSampleField",