+++ lightly_studio_view/src/lib/components/SampleGridItem/SampleGridItem.svelte
@@ -55,43 +55,27 @@
     }
 </script>
 
-<div
-    class="relative select-none"
-    class:sample-selected={$selectedSampleIds.has(sampleId)}
-    {style}
-    data-testid={dataTestId}
-    data-sample-id={sampleId}
-    data-sample-name={dataSampleName}
-    data-index={index}
-    {ondblclick}
-    onclick={handleOnClick}
-    onkeydown={handleKeyDown}
-    aria-label={`View sample: ${dataSampleName}`}
-    role="button"
-    tabindex="0"
->
-    {#if hasMinimumRole(user?.role, 'labeler')}
-        <div class="absolute right-7 top-1 z-10">
-            <SelectableBox
-                onSelect={() => undefined}
-                isSelected={$selectedSampleIds.has(sampleId)}
-            />
-        </div>
-    {/if}
-
+<div class="select-none" {style}>
     <div
         class="relative overflow-hidden rounded-lg"
+        class:grid-item-selected={$selectedSampleIds.has(sampleId)}
         style="width: var(--sample-width); height: var(--sample-height);"
+        data-testid={dataTestId}
+        data-sample-id={sampleId}
+        data-sample-name={dataSampleName}
+        data-index={index}
+        {ondblclick}
+        onclick={handleOnClick}
+        onkeydown={handleKeyDown}
+        aria-label={`View sample: ${dataSampleName}`}
+        role="button"
+        tabindex="0"
     >
+        {#if hasMinimumRole(user?.role, 'labeler') && $selectedSampleIds.has(sampleId)}
+            <div class="pointer-events-none absolute right-2 top-1.5 z-10" inert>
+                <SelectableBox onSelect={() => undefined} isSelected={true} />
+            </div>
+        {/if}
         {@render item()}
     </div>
 </div>
-
-<style>
-    .sample-selected {
-        outline: drop-shadow(1px 1px 1px hsl(var(--primary)))
-            drop-shadow(1px -1px 1px hsl(var(--primary)))
-            drop-shadow(-1px -1px 1px hsl(var(--primary)))
-            drop-shadow(-1px 1px 1px hsl(var(--primary)));
-    }
-</style>