+++ lightly_studio_view/src/lib/components/index.ts
@@ -24,7 +24,6 @@ export { default as SampleImage } from '$lib/components/SampleImage/index.svelte
 export { default as Samples } from '$lib/components/Samples/Samples.svelte';
 export { default as SelectableBox } from '$lib/components/SelectableBox/SelectableBox.svelte';
 export { SettingsDialog } from '$lib/components/Settings/';
-export { default as TagCreateDialog } from '$lib/components/TagCreateDialog/TagCreateDialog.svelte';
 export { default as TagsMenu } from '$lib/components/TagsMenu/TagsMenu.svelte';
 export { default as UserInfo } from '$lib/components/UserInfo/index.svelte';
 export { default as CreateSelectionDialog } from '$lib/components/Selection/CreateSelectionDialog.svelte';
@@ -55,4 +54,5 @@ export { default as VideoDetailsNavigation } from '$lib/components/VideoDetailsN
 export { default as VideoFrameDetails } from '$lib/components/VideoFrameDetails/VideoFrameDetails.svelte';
 export { default as GroupsComponentsMenu } from '$lib/components/GroupsComponentsMenu/GroupsComponentsMenu.svelte';
 export { default as LayoutCard } from '$lib/components/LayoutCard/LayoutCard.svelte';
+export { default as SelectionPill } from '$lib/components/SelectionPill/SelectionPill.svelte';
 export { default as Separator } from '$lib/components/ui/separator/separator.svelte';